+++ src/main/java/com/seq/api/Key.java
@@ -51,9 +51,9 @@ public class Key {
   }
 
   /**
-   * A builder class for querying keys in the ledger.
+   * A builder class for listing keys in the ledger.
    */
-  public static class QueryBuilder extends BaseQueryBuilder<QueryBuilder> {
+  public static class ListBuilder extends BaseQueryBuilder<ListBuilder> {
     /**
      * Executes the query, returning a page of keys that match the query.
      * @param client ledger API connection object